Reworked the simulation loader, mainly moved format: JSON to Properties.
Removed the JSON dependency, no need for it, really. The new
info.properties is still as simple as the old info.json. Also, the main
property is actually used (it just looked for a Simulation subclass
before). However, if the properties file is missing or main is "", the
it still looks for the first Simulation subclass.