QOL changes: getN() changed to getFrameNumber(), frames start at 0.
Simulation.getN() has been changed to Simulation.getFrameNumber(), a
more informative name. However, for elegance and with the goal to keep
simulations simple, the actual variable behind
Simulation.getFrameNumber(), n, is now protected instead of private, so
that the current frame can be accessed simple as "n" in simulations.

Also, frame number (n) starts at 0 instead of 1 now. If you animation
relied on it starting at 1, just increment it in the reset function.