Added Window–>Minimize feature. Rewrapped Console as a JFrame itself.
File–>Open Content Folder and File–>Open Home Folder were remapped to D
instead of M, because M is used by minimize.

An interface, Minimizable, was created which requires the class to
specify a minimize function. This way, when minimize is called, the menu
checks to see if its parent is minimizable, and if it is, minimizes it.

This also resulted in restructuring Console so that it could just
implement JFrame and an inner class was added to handle the dirty work.