Improved format/colors of the home screen.
Red/pink color schme introduced. The bad gridbag layout issues (where
the width of the first column would change a bit when clicked) has been
fixed pretty much by omitting the weightx from the simulationList.

Also added a colored border to the right of the list as the dark border
and changed the border to the left of the info panel to the same color
to make it look like a margin.

Removed a few other uneeded lines.