Implemented more preferences, added print feature, and more logging.
Simulation–>Play shortcut was changed from P to L because P should
really be for print.

The preferences still are not customizable or saved, the default
preferences are run every time. This is all in preparation for adding a
customizable preference system later.

More debug messages have been added to various places in the program.

Commons IO dependency added, it was useful for removing the temp folder
created for printing images. It will surely come in handy at sometime
later, and is a small file size.