Improved temp folder handling/structure. Added shutdown hook to delete.
Made the temp folder handled through the preferences much like the other
preferences, but did not make it a preference itself. It's more of a
pseudo-preference: it is a directory called ".temp" inside of the save
directory.

Also added a shutdown hook to hopefully make sure that the temp folder
is deleted on program termination so that more disk space is not used
up.