Changed subframe classes to have-a JFrame vs be-a JFrame.
Namely Animate, Controls, Preferences, and Resize.

Modifications were required in other parts of the code, such as the
minimize feature which would normally minimize the owner of the menu
(which was-a JFrame, but is now just a regular JFrame. An owner field
was added to menus so that they could have a non-JFrame owner).

Modifications and refactoring was required all around in order to meet
this new format.