Big fat menu overhaul, mainly based around custom annotations.
Mainly annotations MenuItem and MenuHeader on JMenuItem and JMenu,
respectively. This is done in a subclass of Menu, which, after
initialization, has a getMenuBar() method.

Also, JMenuItem actions are able to be handled by a MenuHandler child
class. Every JMenuItem from the Menu subclass invokes a method of the
same name in the MenuHandler subclass.

The Minimizable interface was removed in favor of a much more useful
FrameHolder interface. There are many classes that "have a" frame and
this interface allows for them to be generalized into objects with a
getFrame() method.

Miscellaneous refactoring includes
com.marklalor.javasim.preferences.serializers being renamed to
com.marklalor.javasim.preferences.jsonserializers.