Revamped Image system. Added ImageLayers, alpha background, and more.
The simulation image is now created from many layers of BufferedImages
which are created by ImageLayers: objects with an index, an option to be
temporary (cleared each frame) or permanent (not), and a DrawHandler
which draws onto the ImageLayer.

Also the structure has changed to move more towards Image and away from
Simulation. The Simulation class is becoming cluttered and in the future
more logic should be administered to different classes for more options
for flexibility.