Use JDK/Apple-supplied OS X event listeners. Removed old Apple listeners
This is also done in a way that makes use of reflection so that on
windows, no references/imports or anything have to be made to packages
that do not exist in the installed JDK. As far as I know, this would
cause an error. Still not tested on Windows, however.