Allowed a Menu to have multiple MenuHandlers and split up the handler
Added Google Guava for an easy way to get the MenuHandlers in the same
package as the Menu. I hear goods things about Guava so it should be a
useful ~1MB asset in the future.

Anyways, using reflection a list of Method objects are created for each
of the menu items from whichever class they are found in.