Moving animation-related logic out of the cluttered Simulation class
PlayManager was created, and each Simulation has a PlayManager. From
there, each of the three animating timers have their own listener
classes, and all of the frequency-calculating methods are contained in
FrequencyMonitor (also renamed parts of fields and methods with "hertz"
to "frequency" because I'd rather use the name of the actual property
rather than the name of the unit.